Changed the connection method between the client and server.
I changed the server to the same TCP connection method as the client, and also made sure to keep the unique number sent by the client so that I could write the appropriate content to the log file for each unique number. I will ask questions in the lab class and then write error checks and comments.